Clarified reporting info, made minor changes
I clarified the logging on the server and the client
to improve tracing execution.

I made a few miscellaneous changes to the code for some
of the ordering, everything should work the same as it did.

I updated the documentation a bit

I had to run, so not much testing has been done.